tests/wpt/css-tests/cssom-view-1_dev/xhtml1/htmlelement-offset-width-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSSOM View - 7 - element.offsetWidth detatches correctly</title>
    <link href="mailto:michael@notriddle.com" rel="author" title="Michael Howell" />
    <link href="https://drafts.csswg.org/cssom-view-1/#dom-htmlelement-offsetwidth" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="element.offsetWidth returns 0 when there is no documentElement." name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="myDiv">...</div>
    <div id="log"></div>
    <script>
        test(function() {
            // These asserts need to be in this order to trigger a bug in Servo.
            var el = document.createElement("div");
            el.appendChild(document.createTextNode("..."));
            assert_equals(el.offsetWidth, 0, "new element has offset width not equal to zero");
            assert_not_equals(document.getElementById("myDiv"), 0, "element with content has offset width equal to zero");
            document.documentElement.remove();
            assert_equals(el.offsetWidth, 0, "new element has offset width not equal to zero after removing documentElement");
        });
    </script>


</body></html>
